chore: Update CSS classes, padding values, and form validation in CSS and JS files - Separated JS files for consistency and readability. - Registration process almost completed, missing a few small things here and there. - Added email validation using Sendgrid and some code refactoring. - Started taking advantage of PSQL to store important informations rather than caching them.
