+++ static/js/auth/auth.js
@@ -1,0 +1,6 @@
+import * as google from './_gg_login.js';
+import * as login from './_login.js';
+import * as auth_events from './auth_events.js';
+import * as proxy from './_proxy.js';
+import * as register from './_register.js';
+import * as pvc from './_pvc.js';
