+++ templates/auth/auth.html
@@ -20,7 +20,7 @@
     <link rel="icon" href="../static/favicon_wuwa.png"/>
     <link rel="stylesheet" href="../static/css/output.css"/>
     <script src="https://accounts.google.com/gsi/client" async defer></script>
-    <script src="../../static/js/auth.js"></script>
+    <script type='module' src="../../static/js/auth/auth.js"></script>
 </head>
 <!-- Google tag (gtag.js) -->
 <script async src="https://www.googletagmanager.com/gtag/js?id=G-44JD26YV2G"></script>
@@ -38,7 +38,7 @@
             <path d="M2.146 2.854a.5.5 0 1 1 .708-.708L8 7.293l5.146-5.147a.5.5 0 0 1 .708.708L8.707 8l5.147 5.146a.5.5 0 0 1-.708.708L8 8.707l-5.146 5.147a.5.5 0 0 1-.708-.708L7.293 8z"/>
         </svg>
     </div>
-    <div class="h-auto w-full bg-transparent md:bg-[#262626] z-10">
+    <div class="absolute top-0 h-auto w-full bg-transparent md:bg-[#262626] z-10">
         <div class="flex items-center justify-center p-2">
             <img src="https://shoshin.moe/static/shoshinlogo.png" alt="Shoshin Logo" class="h-16 w-auto md:h-12" />
         </div>
