+++ templates/auth/register.html
@@ -20,7 +20,7 @@
     <link rel="icon" href="../static/favicon_wuwa.png"/>
     <link rel="stylesheet" href="../static/css/output.css"/>
     <script src="https://accounts.google.com/gsi/client" async defer></script>
-    <script src="../../static/js/auth.js"></script>
+    <script type='module' src="../../static/js/auth/auth.js"></script>
 </head>
 <!-- Google tag (gtag.js) -->
 <script async src="https://www.googletagmanager.com/gtag/js?id=G-44JD26YV2G"></script>
@@ -32,6 +32,12 @@
   gtag('config', 'G-44JD26YV2G');
 </script>
 <body class="sans-it">
+    <div id="_sho-register-errorBannerTop" class="hidden absolute top-0 w-full h-auto p-4 bg-red-600 flex items-center justify-center z-30 transition-transform duration-500 ease-out transform -translate-y-full">
+        <p id="_sho-register-errorBannerTop-text" class="text-black text-lg">Email or password is incorrect.</p>
+        <svg id="_sho-register-errorBannerTop-close" xmlns="http://www.w3.org/2000/svg" width="20" height="20" fill="black" class="absolute right-2 bi bi-x-lg hover:cursor-pointer" viewBox="0 0 16 16">
+            <path d="M2.146 2.854a.5.5 0 1 1 .708-.708L8 7.293l5.146-5.147a.5.5 0 0 1 .708.708L8.707 8l5.147 5.146a.5.5 0 0 1-.708.708L8 8.707l-5.146 5.147a.5.5 0 0 1-.708-.708L7.293 8z"/>
+        </svg>
+    </div>
     <div class="absolute top-0 h-auto w-full bg-transparent md:bg-[#262626] z-10">
         <div class="flex items-center justify-center p-2">
             <a href="https://beta.shoshin.moe" class="hover:cursor-pointer">
@@ -45,10 +51,10 @@
                 <div class="absolute -left-44 -top-5 z-0">
                     <img src="https://shoshin.moe/static/register.png" alt="Login Cutie" class="w-52 hidden lg:block" />
                 </div>
-                <div class="rounded-sm bg-[#333333] p-8 z-30 space-y-4">
-                    <div class="flex flex-col gap-2">
-                        <p class="mx-auto text-2xl text-white md:text-3xl">Create an account!</p>
-                        <p class="mx-auto text-xs text-gray-300 mb-4">By registering, you accept our 
+                <div id="_sho-register-form" class="rounded-sm bg-[#333333] p-8 z-30 space-y-4">
+                    <div id="_sho-register-TopFormDiv" class="flex flex-col gap-2">
+                        <p id="_sho-register-formHeaderOne" class="mx-auto text-2xl text-white md:text-3xl">Create an account!</p>
+                        <p id="_sho-register-formHeaderTwo" class="mx-auto text-xs text-gray-300 mb-4">By registering, you accept our 
                             <a class="text-blue-500 hover:text-blue-700 hover:underline" href="https://shoshin.moe/terms">Terms of Service</a> 
                             and 
                             <a class="text-blue-500 hover:text-blue-700 hover:underline" href="https://shoshin.moe/privacy">Privacy Policy</a>.
